move register.components out of register folder
