+++ src/app/modules/rc-auth/components/register-teacher/register-teacher.component.html
