+++ src/app/modules/rc-auth/rc-auth-routing.module.ts
@@ -3,8 +3,8 @@ import {RouterModule, Routes} from '@angular/router';
 import {LoginComponent} from "./components/login/login.component";
 import {RegisterComponent} from "./components/register/register.component";
 import {AuthComponent} from "./components/auth/auth.component";
-import {RegisterTeacherComponent} from "./components/register/register-teacher/register-teacher.component";
-import {RegisterStudentComponent} from "./components/register/register-student/register-student.component";
+import {RegisterTeacherComponent} from "./components/register-teacher/register-teacher.component";
+import {RegisterStudentComponent} from "./components/register-student/register-student.component";
 
 const routes: Routes = [
