+++ src/app/modules/rc-auth/rc-auth.module.ts
@@ -4,8 +4,8 @@ import {LoginComponent} from './components/login/login.component';
 import {RegisterComponent} from './components/register/register.component'
 import {AuthComponent} from './components/auth/auth.component';
 import {RcAuthRoutingModule} from "./rc-auth-routing.module";
-import {RegisterTeacherComponent} from './components/register/register-teacher/register-teacher.component';
-import {RegisterStudentComponent} from './components/register/register-student/register-student.component';
+import {RegisterTeacherComponent} from './components/register-teacher/register-teacher.component';
+import {RegisterStudentComponent} from './components/register-student/register-student.component';
 import {RcSharedModule} from "../rc-shared.module";
